[refactor] move prober API and config impl out of main.go
